[Bug/MiddleEnd]: Fix the float type handling and add a new test-case.

+++ tests/backend/aarch64-fp-arithmetic-2.c
@@ -1,18 +1,18 @@
 // RUN: AArch64
 
-// FUNC-DECL: int test(double,double)
+// FUNC-DECL: int test(float,float)
 // TEST-CASE: test(0, 1.0) -> 1 + -1 + 0 + 0
 // TEST-CASE: test(0.5, 0.5) -> 1 + 0 + 0 + 1
 // TEST-CASE: test(0.8, 0.3) -> 1 + 0 + 0 + 2
 // TEST-CASE: test(-1.3, -1.8) -> -3 + 0 + 2 + 0
 
 
-int test_add(double a, double b) { return a + b; }
-int test_sub(double a, double b) { return a - b; }
-int test_mul(double a, double b) { return a * b; }
-int test_div(double a, double b) { return a / b; }
+int test_add(float a, float b) { return a + b; }
+int test_sub(float a, float b) { return a - b; }
+int test_mul(float a, float b) { return a * b; }
+int test_div(float a, float b) { return a / b; }
 
-int test(double a, double b)
+int test(float a, float b)
 {
     int res = 0;
